services/auth-system/Django-Allauth/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
python-dotenv
sqlparse
django-allauth
requests
oauthlib
PyJWT
cryptography
djangorestframework-simplejwt
drf-yasg
gunicorn
whitenoise
Pillow
aio-pika
channels
uvicorn
pika
psycopg2-binary